simplesim initial commit with changes as it's own directory inside of sst elements (instead of SimpleElement)
