+++ src/sst/elements/simpleSimulation/Makefile.am
@@ -1,0 +1,26 @@
+# -*- Makefile -*-
+#
+#
+
+AM_CPPFLAGS = \
+	$(BOOST_CPPFLAGS) \
+	$(MPI_CPPFLAGS) 
+
+compdir = $(pkglibdir)
+comp_LTLIBRARIES = libsimpleSimulation.la
+libsimpleSimulation_la_SOURCES = \
+	simpleSimulation.cc \
+	simpleCarWash.h \
+	simpleCarWash.cc
+	
+EXTRA_DIST = \
+    README \
+    tests/test_simpleCarWash.py 
+
+libsimpleSimulation_la_LDFLAGS = -module -avoid-version
+
+##########################################################################
+##########################################################################
+##########################################################################
+
+
